# Conflicts:
#	core/camerafixture/sceneFeaturePoint.cpp
#	core/math/sparseMatrix.cpp

+++ core/camerafixture/sceneFeaturePoint.cpp
@@ -87,14 +87,14 @@ int SceneFeaturePoint::ensureDistortedObservations(bool distorted)
     return toReturn;
 }
 
-Vector3dd SceneFeaturePoint::triangulate(bool use__, uint32_t mask)
+Vector3dd SceneFeaturePoint::triangulate(bool use__, uint64_t mask)
 {
     MulticameraTriangulator mct;
     int id = 0;
     if (use__)
     {
         for (auto& obs: observations__)
-            if (mask & (1 << id))
+            if (mask & (1llu << id))
                 mct.addCamera(obs.first.u->getMMatrix(obs.first.v), obs.second.observation);
     }
     else
@@ -102,7 +102,7 @@ Vector3dd SceneFeaturePoint::triangulate(bool use__, uint32_t mask)
         for (auto& obs : observations)
         {
             CORE_ASSERT_TRUE_S(obs.second.cameraFixture != NULL);
-            if (mask & (1 << id))
+            if (mask & (1llu << id))
                 mct.addCamera(obs.second.cameraFixture->getMMatrix(obs.second.camera), obs.second.observation);
         }
     }
